Started converting to new template function
Also started working up the code for a Config struct and a relevant
builder. I might switch to using `derive_builder` and maybe even
`derive_default`? But for now keep plugging away.

Next I need to add filters and functions to the new implementation, and
then after that can refactor them a bit.