Adjusted some visibility and added some docs
Documenting this is kind of awkward and cumbersome. The main important
parts are really the examples and such. I suppose I could go with my
initial instinct and *not* make the `model` module public. I'd still
want to put comments in there, but at that point it's mostly for
development.

At the end of the day, this is a template engine that is meant to be
used, but not modified internally, right?